+++ core-ng/src/test/java/core/framework/internal/http/FallbackDNSCacheTest.java
@@ -5,6 +5,9 @@ import org.junit.jupiter.api.Test;
 
 import java.net.InetAddress;
 import java.net.UnknownHostException;
+import java.time.Clock;
+import java.time.Instant;
+import java.time.ZoneId;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -15,10 +18,12 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
  */
 class FallbackDNSCacheTest {
     private FallbackDNSCache dns;
+    private Clock clock;
 
     @BeforeEach
     void createFallbackDNSCache() {
-        dns = new FallbackDNSCache();
+        clock = Clock.fixed(Instant.now(), ZoneId.systemDefault());
+        dns = new FallbackDNSCache(clock);
     }
 
     @Test
@@ -30,12 +35,17 @@ class FallbackDNSCacheTest {
     @Test
     void lookupWithInvalidDomain() {
         assertThatThrownBy(() -> dns.lookup("www.invalid"))
-                .isInstanceOf(UnknownHostException.class);
+            .isInstanceOf(UnknownHostException.class);
     }
 
     @Test
     void lookupWithFallback() throws UnknownHostException {
-        dns.cache.put("test.invalid", List.of(InetAddress.getLocalHost()));
+        dns.cache.put("test.invalid", new FallbackDNSCache.Entry(List.of(InetAddress.getLocalHost()), clock.millis() + 1_000));
         assertThat(dns.lookup("test.invalid")).isNotNull();
+
+        dns.cache.put("test.invalid", new FallbackDNSCache.Entry(List.of(InetAddress.getLocalHost()), clock.millis() - 1_000));
+        assertThatThrownBy(() -> dns.lookup("test.invalid"))
+            .isInstanceOf(UnknownHostException.class);
+        assertThat(dns.cache).doesNotContainKey("test.invalid");
     }
 }